requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
transformers==4.31.0
accelerate
nltk
ipython
tabulate
python-Levenshtein
